requirements.txt
numpy
pandas
scikit-learn
tensorflow
xgboost
lightgbm
mlflow
evidently
fastapi
uvicorn
mangum
boto3
awscli
streamlit
apache-airflow
pytest
flake8
requests
